multi-switchport-descriptions
I had to update multiple port descriptions on a switch so after bunch of trial and error I was able to come up with this script. It uses a csv file where you list the ports and their descriptions.  Initially for the input you have to provide the device(switch) ID to the script and then it will iterate through the csv and push out the changes.  It is a work in progress and I will most likely add different versions of it.  One of the issue I ran into was that there is no merging of the config so my very first script was running but each line in the csv was clearing the previous commands.  Which made the whole process complicated.